Reorganize modules/ by project (BIG CHANGE WARNING)
Files in modules/sofa/component/ were organised only according to the
namespace hierarchy, this commit reorganizes files by project (in the
cmake sense, like, checkboxes in cmake-gui).

E.g. files in
  modules/sofa/component/SofaBaseAnimation/
are moved to
  modules/BaseAnimation/

Moreover, the "Sofa" prefix is dropped, and the directory structure in
each module is completely flattened, unless a module contains many files.

On the pragmatic side, this changes what you put in:
- include directives: <sofa/component/foo/bar/baz.h> -> <ModuleName/baz.h>
- linker dependencies: SofaBaseCollision -> BaseCollision

But fear not, this commit also provides a script to fix those:
$ ./scripts/fix-modules-includes.sh path/to/your/plugin
This should take care of most changes. Obviously, this will modify your
files, so needless to say, you should save your work beforehand.
It Works on My Machine. Hopefully, it will not wreak havoc on yours.

Oh, also, you should completely rebuild SOFA, because there is no way on
Earth the cmake scripts can handle this.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/sofa/trunk/Sofa@10771 b7f1a160-662a-0410-a6d8-bea165342f16