push 3ba7512..863f934
* commit 863f9341f1f952a9d9fa3f864a75df2dae1a33f0
Move tests resources back to tests directories

They were moved into share/ in a previous commit, but this was not a
smart move (hehe). As it turns out, absolute paths are perfectly OK for
test resources, and defining a path as a macro in a CMakeLists.txt is
perfectly robust provided it is a string literal.

   34 files changed, 747 insertions(+), 742 deletions(-)



git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/sofa/trunk/Sofa@10885 b7f1a160-662a-0410-a6d8-bea165342f16