push 6860d1c..f40f4df
* commit f40f4df59f12f024b8f1a078ed26916ff9cda2c2
ObjectFactory: use smart pointers

This commit replaces raw pointers with instances of boost::shared_ptr in
the ObjectFactory (for instances of Creator and ClassEntry). As a
consequence, this changes the prototype of getEntry, addAlias, AddAlias,
resetAlias and ResetAlias.

|  9 files changed, 75 insertions(+), 81 deletions(-)

* commit cdb1252ec45d0e5d634221ec3d9a2c17379d3bf7
ObjectFactory: clean up, remove unused code

|  2 files changed, 0 insertions(+), 58 deletions(-)

* commit d303035475e576f18bf3bca6fe74124aacc54341
ObjectFactory: remove ClassEntryList, CreatorList

ClassEntry and Creators are stored both in a map and in a list, for no
obvious reason (I guess, to avoid breaking the interface). This commit
gets rid of this redundancy, by using only maps and adapting the code
that used the list versions instead of the maps.

   7 files changed, 25 insertions(+), 32 deletions(-)



git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/sofa/trunk/Sofa@10953 b7f1a160-662a-0410-a6d8-bea165342f16